backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
sqlite3  # Built-in with Python
lxml==4.9.3
beautifulsoup4==4.12.2
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
gunicorn==21.2.0  # För produktion
watchdog==3.0.0  # För filövervakning
click==8.1.7
Werkzeug==2.3.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
Jinja2==3.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
urllib3==2.0.4
